Support having multiple stat lists per spec
This is so tank specs can list separate specs for survivability and
damage output.

Right now the LDB text will always just use the first stat from the
first stat list. It would be nice to support picking a stat type (and
then showing just that type in the tooltip), but that requires saved
data and some way to pick (e.g. a right-click menu), which is more
complicated.